More centered, spacing still weird, code tightened etc.
- By "spacing still weird" I mean that sometimes the letters
appear farther and sometimes closer. For example, if I make the
distance any less, when there are 3 qubits the labels will be
inside the qubits. Other times there's a lot of room between
the qubits and the labels. When there's just one qubit, the letter
"A" is at the top of the visualizer. I'm not sure what's going
on with arrangeRadius that it's doing this, so I'm not sure how
to fix it.